Extract common logic from CrashModule and MissingCrashModule
