+++ src/main/kotlin/io/github/mojira/arisa/modules/MissingCrashModule.kt
@@ -2,12 +2,10 @@ package io.github.mojira.arisa.modules
 
 import arrow.core.Either
 import arrow.core.extensions.fx
-import io.github.mojira.arisa.domain.Attachment
 import io.github.mojira.arisa.domain.CommentOptions
 import io.github.mojira.arisa.domain.Issue
-import me.urielsalis.mccrashlib.Crash
+import io.github.mojira.arisa.infrastructure.AttachmentUtils
 import me.urielsalis.mccrashlib.CrashReader
-import me.urielsalis.mccrashlib.parser.ParserError
 import java.time.Instant
 
 class MissingCrashModule(
@@ -22,20 +20,7 @@ class MissingCrashModule(
             assertNull(priority).bind()
             assertContains(description, "crash").bind()
 
-            val textDocuments = attachments
-                .filter { isCrashAttachment(it.name) }
-                .map(::fetchAttachment)
-                .toMutableList()
-            // also add description, so it's searched for crash reports
-            textDocuments.add(TextDocument({ description ?: "" }, created))
-
-            val crashes = textDocuments
-                .asSequence()
-                .map { processCrash(it) }
-                .filter { it.second.isRight() }
-                .map { extractCrash(it) }
-                .filter { it.second is Crash.Minecraft || it.second is Crash.Java }
-                .toList()
+            val crashes = AttachmentUtils(crashReportExtensions, crashReader).extractCrashesFromAttachments(issue)
 
             assertEmpty(crashes).bind()
 
@@ -43,26 +28,4 @@ class MissingCrashModule(
             addComment(CommentOptions(message))
         }
     }
-
-    private fun extractCrash(it: Pair<TextDocument, Either<ParserError, Crash>>) =
-        it.first to (it.second as Either.Right<Crash>).b
-
-    private fun processCrash(it: TextDocument) = it to crashReader.processCrash(it.getContent().lines())
-
-    private fun isCrashAttachment(fileName: String) =
-        crashReportExtensions.any { it == fileName.substring(fileName.lastIndexOf(".") + 1) }
-
-    private fun fetchAttachment(attachment: Attachment): TextDocument {
-        val getText = {
-            val data = attachment.getContent()
-            String(data)
-        }
-
-        return TextDocument(getText, attachment.created)
-    }
-
-    data class TextDocument(
-        val getContent: () -> String,
-        val created: Instant
-    )
 }
